refactor: move API cache modules into cache package
