+++ frontend/src/api/currency/hooks.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { fetchCurrencies } from '@/api/currency/requests';
-import { currencyKeys } from '@/api/queryKeys';
+import { currencyKeys } from '@/api/cache/queryKeys';
 
 /**
  * Reads currencies with session-long caching because ISO 4217 metadata is effectively static
