+++ frontend/src/api/transactions/index.ts
@@ -23,7 +23,7 @@ export {
 
 export {
   invalidateTransactionAccountData,
-} from '@/api/transactions/cache';
+} from '@/api/cache/transactions';
 
 export {
   useCreateTransaction,
